refactor: reorganize internal packages and move domain to internal
- Move domain package from project root to internal/domain for better encapsulation
- Rename internal/server to internal/http to better reflect HTTP-specific functionality
- Update all import paths to use new internal package locations
- Use import aliases (httpserver) to avoid naming conflicts with standard library
- Remove old api.yaml file and rename to openapi.yaml for consistency
- Add test documentation comments for better code clarity
- Clean up empty directories and minor typo fixes in README
