+++ features/driver/http/server.go
@@ -2,10 +2,10 @@ package http
 
 import (
 	"github.com/sirockin/cucumber-screenplay-go/features/driver"
-	"github.com/sirockin/cucumber-screenplay-go/internal/server"
+	httpserver "github.com/sirockin/cucumber-screenplay-go/internal/http"
 )
 
 // NewServer creates a new HTTP server wrapping the given ApplicationDriver
-func NewServer(app driver.ApplicationDriver) *server.Server {
-	return server.NewServer(app)
+func NewServer(app driver.ApplicationDriver) *httpserver.Server {
+	return httpserver.NewServer(app)
 }
